requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
mysqlclient==2.1.1
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
